+++ src/GSvar/ProcessedSampleWidget.cpp
@@ -196,6 +196,7 @@ void ProcessedSampleWidget::updateGUI()
 	ancestry_details << "SAS (South asian): " + db.getValue("SELECT score_sas FROM processed_sample_ancestry WHERE processed_sample_id="+ps_id_, true).toString();
 	ancestry_details << "EAS (East asian): " + db.getValue("SELECT score_eas FROM processed_sample_ancestry WHERE processed_sample_id="+ps_id_, true).toString();
 	ui_->ancestry->setToolTip(ancestry_details.join("\n"));
+	ui_->urgent->setText(ps_data.urgent ? "<font color=red>yes</font>" : "");
 
 	//#### sample details ####
 	QString s_id = db.getValue("SELECT sample_id FROM processed_sample WHERE id='" + ps_id_ + "'").toString();