+++ src/GSvarServer/ServerController.cpp
@@ -901,7 +901,7 @@ HttpResponse ServerController::annotateVariant(const HttpRequest& request)
     process.setProcessChannelMode(QProcess::MergedChannels);
     QString an_vep_out = Helper::tempFileName(".vcf");
     Log::info("Running megSAP >> an_vep.php: " + an_vep_out);
-	process.start("php", QStringList() << PipelineSettings::rootDir() + "/src/NGS/an_vep.php" << "-in" << input_vcf << "-out" << an_vep_out);
+	process.start("php", QStringList() << PipelineSettings::rootDir() + "/src/Tools/an_vep.php" << "-in" << input_vcf << "-out" << an_vep_out);
     bool success = process.waitForFinished(-1);
     Log::error("Exit code = " + QString::number(process.exitCode()));
     if (!success || process.exitCode()>0)
@@ -912,7 +912,7 @@ HttpResponse ServerController::annotateVariant(const HttpRequest& request)
 
     QString vcf2gsvar_out = Helper::tempFileName(".GSvar");
     Log::info("Running megSAP >> vcf2gsvar.php: " + vcf2gsvar_out);
-	process.start("php", QStringList() << PipelineSettings::rootDir() + "/src/NGS/vcf2gsvar.php" << "-in" << an_vep_out << "-out" << vcf2gsvar_out);
+	process.start("php", QStringList() << PipelineSettings::rootDir() + "/src/Tools/vcf2gsvar.php" << "-in" << an_vep_out << "-out" << vcf2gsvar_out);
     success = process.waitForFinished(-1);
     if (!success || process.exitCode()>0)
     {