+++ doc/GSvar/changelog.md
@@ -1,5 +1,17 @@
 # GSvar change log
 
+## 2025_01-22 (06.02.25)
+
+- sequencing run tab: added settings dialog for NovaSeqX sample sheet creation
+- variant context menu: added new PubMed search string
+
+## 2025_01-9 (29.01.25)
+
+- general: added 'urgent' flag to processed samples (run tab, processed sample tab, batch import)
+- general: added context menu to tab bar
+- small variants: added column context menu to allow faster column-specific filtering
+- sample search: sample name now supports asterisk at the start/end
+
 ## 2024_11-86 (20.01.25)
 
 - general: renamed class 'R*' to 'R' (changed in HerediCare and HerediVar)